feat(paper): migrate account money to DECIMAL, and make the schema gap visible
净值恒等式 equity == cash + shares × price 要能**零容差**成立，而 double 结构上做不到。
理由用一条断言写死（MoneyTest）：

    100 股 × 10.1234 元 = 1012.34，加上现金 8999.87 应当恰好等于 10012.21
    用 double 算，这个和是 10012.210000000001

一次买入差 1e-12，几百笔之后"净值怎么少了几分钱"就再也说不清了 —— 而净值曲线、
回撤、对账、"净值什么时候真的变了"全都建在这条等式上。

改了什么：

- `PaperAccount` / `PaperTrade` 的金额字段 → `BigDecimal` + `precision/scale`
  （金额/净值 2 位、价格/股数 4 位，与 `Money` / Python `MoneyPolicy` 同一口径）
- `PaperTradingService` 的结算路径全部改为定点运算：整手取整（`divide(...,0,DOWN)`
  而不是 `Math.floor`）、佣金 5 元下限、印花税、净值与最高水位、收益率；
  `PaperAccountResponse` / `PaperTradeResponse` 同步（JSON 仍是数字，调用方无感）
- `ObjectiveFactKeys.format` 增加 BigDecimal 分支：事实通道的值格式仍由**一处**决定
  （≤3 位小数、整数归一）。少了这一支，同一个量会写成 "-25.0000" 与 "-25" 两种形式 ——
  那正是"值没变、取代链却多一条"的来源

`ddl-auto=update` **只加列、不改列类型**，所以老库里的列还是 DOUBLE：
- `docs/superpowers/specs/2026-09-18-paper-money-decimal.sql`：可执行的一次性脚本
  + 迁移后的自检查询
- `config/MoneySchemaCheck`：启动时用 JDBC 元数据查一遍（H2 与 MySQL 同一段代码），
  不一致就 WARN 并把 ALTER 打出来。只告警不中断 —— 但绝不静默：
  "以为迁了、其实没迁"的钱包比没迁更危险。配套测试反向验证"检查真的读得到列"
  （否则它的"没问题"是假绿）。

顺带踩到并修掉的两件事：
- `mvnw -o compile` 出现 "Nothing to compile" 的陈旧增量状态，把编译错误藏了一轮 ——
  这次迁移用 `clean test-compile` 复核。
- 用 PowerShell 文本 cmdlet 往返源文件会把中文变成乱码（这个仓库中文很重），
  已回滚重做；源码一律只用编辑工具改。

测试：Java 182 passed（新增 schema 检查 2、钱的迁移理由 1；结算与报告用例改为
BigDecimal 值比较，避免把"标度不同"报成"金额不同"）；Python 630 passed / 29 skipped；
真实行情端到端那条也重跑通过（报告渲染走的就是 BigDecimal 账户）。

+++ src/main/java/com/happyericsix/stocktracker/dto/PaperAccountResponse.java
@@ -2,17 +2,22 @@ package com.happyericsix.stocktracker.dto;
 
 import com.happyericsix.stocktracker.entity.PaperAccount;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
+/**
+ * 账户的读取视图。金额一律 {@link BigDecimal}：账里的精度不该在展示层被换成 double
+ * （JSON 里仍然是数字，前端与调用方感知不到差别）。
+ */
 public class PaperAccountResponse {
     private Long id;
-    private Double initialCapital;
-    private Double cash;
-    private Double shares;
-    private Double avgCost;
-    private Double equity;
-    private Double highWatermark;
-    private Double lastPrice;
+    private BigDecimal initialCapital;
+    private BigDecimal cash;
+    private BigDecimal shares;
+    private BigDecimal avgCost;
+    private BigDecimal equity;
+    private BigDecimal highWatermark;
+    private BigDecimal lastPrice;
     private String lastSignal;
     private LocalDateTime lastEvalAt;
     private LocalDateTime createdAt;
@@ -20,9 +25,9 @@ public class PaperAccountResponse {
 
     public PaperAccountResponse() {}
 
-    public PaperAccountResponse(Long id, Double initialCapital, Double cash, Double shares,
-                                Double avgCost, Double equity, Double highWatermark,
-                                Double lastPrice, String lastSignal, LocalDateTime lastEvalAt,
+    public PaperAccountResponse(Long id, BigDecimal initialCapital, BigDecimal cash, BigDecimal shares,
+                                BigDecimal avgCost, BigDecimal equity, BigDecimal highWatermark,
+                                BigDecimal lastPrice, String lastSignal, LocalDateTime lastEvalAt,
                                 LocalDateTime createdAt, LocalDateTime updatedAt) {
         this.id = id;
         this.initialCapital = initialCapital;
@@ -57,20 +62,20 @@ public class PaperAccountResponse {
 
     public Long getId() { return id; }
     public void setId(Long id) { this.id = id; }
-    public Double getInitialCapital() { return initialCapital; }
-    public void setInitialCapital(Double initialCapital) { this.initialCapital = initialCapital; }
-    public Double getCash() { return cash; }
-    public void setCash(Double cash) { this.cash = cash; }
-    public Double getShares() { return shares; }
-    public void setShares(Double shares) { this.shares = shares; }
-    public Double getAvgCost() { return avgCost; }
-    public void setAvgCost(Double avgCost) { this.avgCost = avgCost; }
-    public Double getEquity() { return equity; }
-    public void setEquity(Double equity) { this.equity = equity; }
-    public Double getHighWatermark() { return highWatermark; }
-    public void setHighWatermark(Double highWatermark) { this.highWatermark = highWatermark; }
-    public Double getLastPrice() { return lastPrice; }
-    public void setLastPrice(Double lastPrice) { this.lastPrice = lastPrice; }
+    public BigDecimal getInitialCapital() { return initialCapital; }
+    public void setInitialCapital(BigDecimal initialCapital) { this.initialCapital = initialCapital; }
+    public BigDecimal getCash() { return cash; }
+    public void setCash(BigDecimal cash) { this.cash = cash; }
+    public BigDecimal getShares() { return shares; }
+    public void setShares(BigDecimal shares) { this.shares = shares; }
+    public BigDecimal getAvgCost() { return avgCost; }
+    public void setAvgCost(BigDecimal avgCost) { this.avgCost = avgCost; }
+    public BigDecimal getEquity() { return equity; }
+    public void setEquity(BigDecimal equity) { this.equity = equity; }
+    public BigDecimal getHighWatermark() { return highWatermark; }
+    public void setHighWatermark(BigDecimal highWatermark) { this.highWatermark = highWatermark; }
+    public BigDecimal getLastPrice() { return lastPrice; }
+    public void setLastPrice(BigDecimal lastPrice) { this.lastPrice = lastPrice; }
     public String getLastSignal() { return lastSignal; }
     public void setLastSignal(String lastSignal) { this.lastSignal = lastSignal; }
     public LocalDateTime getLastEvalAt() { return lastEvalAt; }
