+++ src/main/java/com/happyericsix/stocktracker/dto/PaperTradeResponse.java
@@ -2,6 +2,7 @@ package com.happyericsix.stocktracker.dto;
 
 import com.happyericsix.stocktracker.entity.PaperTrade;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -9,9 +10,9 @@ public class PaperTradeResponse {
     private LocalDate tradeDate;
     private String symbol;
     private String side;
-    private Double price;
-    private Double shares;
-    private Double amount;
+    private BigDecimal price;
+    private BigDecimal shares;
+    private BigDecimal amount;
     private String reason;
     private LocalDateTime createdAt;
     /**
@@ -26,7 +27,7 @@ public class PaperTradeResponse {
     public PaperTradeResponse() {}
 
     public PaperTradeResponse(LocalDate tradeDate, String symbol, String side,
-                              Double price, Double shares, Double amount,
+                              BigDecimal price, BigDecimal shares, BigDecimal amount,
                               String reason, LocalDateTime createdAt, Long traceId) {
         this.tradeDate = tradeDate;
         this.symbol = symbol;
@@ -62,12 +63,12 @@ public class PaperTradeResponse {
     public void setSymbol(String symbol) { this.symbol = symbol; }
     public String getSide() { return side; }
     public void setSide(String side) { this.side = side; }
-    public Double getPrice() { return price; }
-    public void setPrice(Double price) { this.price = price; }
-    public Double getShares() { return shares; }
-    public void setShares(Double shares) { this.shares = shares; }
-    public Double getAmount() { return amount; }
-    public void setAmount(Double amount) { this.amount = amount; }
+    public BigDecimal getPrice() { return price; }
+    public void setPrice(BigDecimal price) { this.price = price; }
+    public BigDecimal getShares() { return shares; }
+    public void setShares(BigDecimal shares) { this.shares = shares; }
+    public BigDecimal getAmount() { return amount; }
+    public void setAmount(BigDecimal amount) { this.amount = amount; }
     public String getReason() { return reason; }
     public void setReason(String reason) { this.reason = reason; }
     public LocalDateTime getCreatedAt() { return createdAt; }
